+++ logprep/processor/base/rule.py
@@ -91,7 +91,6 @@ Further details can be found in the section for processors.
     ]
 """
 
-from functools import partial
 import json
 from os.path import basename, splitext
 from typing import List, Set, Optional, Dict
@@ -106,7 +105,6 @@ from logprep.filter.lucene_filter import LuceneFilter
 from logprep.processor.base.exceptions import InvalidRuleDefinitionError
 from logprep.util.json_handling import is_json
 from logprep.util.helper import camel_to_snake
-from logprep.util.validators import min_len_validator
 
 yaml = YAML(typ="safe", pure=True)
 
@@ -143,6 +141,18 @@ class Rule:
             factory=list,
         )
         """Custom tests for this rule."""
+        tag_on_failure: list = field(
+            validator=[
+                validators.instance_of(list),
+                validators.deep_iterable(member_validator=validators.instance_of(str)),
+            ],
+            factory=list,
+            converter=lambda x: list(set(x)),
+        )
+        """A list of tags which will be appended to the event on non critical errors,
+        defaults to :code:`["_<rule_type>_failure"]`.
+        Is currently only used by the Dissector and FieldManager.
+        """
 
     @define(kw_only=True)
     class RuleMetrics(Metric):
@@ -161,11 +171,13 @@ class Rule:
             self._mean_processing_time = new_avg
             self._mean_processing_time_sample_counter = new_sample_counter
 
-    special_field_types = ["regex_fields", "sigma_fields", "ip_fields", "tests"]
+    special_field_types = ["regex_fields", "sigma_fields", "ip_fields", "tests", "tag_on_failure"]
 
     def __init__(self, filter_rule: FilterExpression, config: Config):
         if not isinstance(config, self.Config):
             raise InvalidRuleDefinitionError("config is not a Config class")
+        if not config.tag_on_failure:
+            config.tag_on_failure = [f"_{self.rule_type}_failure"]
         self.__class__.__hash__ = Rule.__hash__
         self.filter_str = str(filter_rule)
         self._filter = filter_rule
@@ -188,6 +200,10 @@ class Rule:
     @property
     def tests(self) -> list:
         return self._config.tests
+
+    @property
+    def failure_tags(self):
+        return self._config.tag_on_failure
 
     # pylint: enable=C0111
 
@@ -220,12 +236,12 @@ class Rule:
     def _create_from_dict(cls, rule: dict) -> "Rule":
         cls.normalize_rule_dict(rule)
         filter_expression = Rule._create_filter_expression(rule)
-        rule_type = camel_to_snake(cls.__name__.replace("Rule", ""))
-        if not rule_type:
-            rule_type = "rule"
-        config = rule.get(rule_type)
+        cls.rule_type = camel_to_snake(cls.__name__.replace("Rule", ""))
+        if not cls.rule_type:
+            cls.rule_type = "rule"
+        config = rule.get(cls.rule_type)
         if config is None:
-            raise InvalidRuleDefinitionError(f"config not under key {rule_type}")
+            raise InvalidRuleDefinitionError(f"config not under key {cls.rule_type}")
         if not isinstance(config, dict):
             raise InvalidRuleDefinitionError("config is not a dict")
         config.update({"description": rule.get("description", "")})
@@ -274,49 +290,3 @@ class Rule:
                 raise ValueError
 
         return special_fields
-
-
-class SourceTargetRule(Rule):
-    """Interface for a simple Rule with source_fields and target_field"""
-
-    @define(kw_only=True)
-    class Config(Rule.Config):
-        """Config for SimpleSourceTargetRule"""
-
-        source_fields: list = field(
-            validator=[
-                validators.instance_of(list),
-                validators.deep_iterable(member_validator=validators.instance_of(str)),
-                partial(min_len_validator, min_length=1),
-            ]
-        )
-        """The fields from where to get the values which should be processed"""
-        target_field: str = field(validator=validators.instance_of(str))
-        """The field where to write the processed values to"""
-        delete_source_fields: str = field(validator=validators.instance_of(bool), default=False)
-        """Whether to delete all the source fields or not. Defaults to :code:`False`"""
-        overwrite_target: str = field(validator=validators.instance_of(bool), default=False)
-        """Overwrite the target field value if exists. Defaults to :code:`False`"""
-
-    # pylint: disable=missing-function-docstring
-    @property
-    def delete_source_fields(self):
-        if hasattr(self, "_config"):
-            return self._config.delete_source_fields
-        return False
-
-    @property
-    def source_fields(self):
-        if hasattr(self, "_config"):
-            return self._config.source_fields
-        return []
-
-    @property
-    def target_field(self):
-        return self._config.target_field
-
-    @property
-    def overwrite_target(self):
-        return self._config.overwrite_target
-
-    # pylint: enable=missing-function-docstring